requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
pyyaml
setuptools
#pickle
#yaml